requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
firebase-admin==6.2.0
google-generativeai==0.3.2
PyPDF2==3.0.1
PyMuPDF==1.23.14
Pillow==10.1.0
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
